+++ src/main/java/com/example/muaring/domain/music/service/MusicService.java
@@ -1,21 +1,13 @@
-package com.example.muaring.domain.social.service;
+package com.example.muaring.domain.music.service;
 
-import com.example.muaring.domain.group.entity.Group;
-import com.example.muaring.domain.group.repository.GroupRepository;
-import com.example.muaring.domain.member.entity.Member;
-import com.example.muaring.domain.member.repository.MemberRepository;
-import com.example.muaring.domain.music.dto.MusicHistoryDTO;
 import com.example.muaring.domain.music.dto.MusicRequestDTO;
 import com.example.muaring.domain.music.dto.SpotifyTrackDTO;
 import com.example.muaring.domain.music.entity.Music;
 import com.example.muaring.domain.music.exception.MusicErrorCode;
 import com.example.muaring.domain.music.exception.MusicException;
-import com.example.muaring.domain.music.response.SpotifyTrackDetailResponse;
-import com.example.muaring.domain.music.service.SpotifyAuthService;
-import com.example.muaring.domain.social.repository.MusicPostRepository;
 import com.example.muaring.domain.music.repository.MusicRepository;
 import com.example.muaring.domain.music.response.SpotifySearchResponse;
-import com.example.muaring.domain.social.entity.MusicPost;
+import com.example.muaring.domain.music.response.SpotifyTrackDetailResponse;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
@@ -27,29 +19,19 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-public class MusicPostService {
+public class MusicService {
 
     private final MusicRepository musicRepository;
-    private final GroupRepository groupRepository;
-    private final MemberRepository memberRepository;
-    private final MusicPostRepository musicPostRepository;
     private final SpotifyAuthService spotifyAuthService;
     private final WebClient webClient;
 
-    public MusicPostService(
+    public MusicService(
             MusicRepository musicRepository,
-            GroupRepository groupRepository,
-            MemberRepository memberRepository,
-            MusicPostRepository musicPostRepository,
-            @Qualifier("spotifyApiWebClient") WebClient webClient,
-            SpotifyAuthService spotifyAuthService
-    ) {
+            SpotifyAuthService spotifyAuthService,
+            @Qualifier("spotifyApiWebClient") WebClient webClient) {
         this.musicRepository = musicRepository;
-        this.groupRepository = groupRepository;
-        this.memberRepository = memberRepository;
-        this.musicPostRepository = musicPostRepository;
-        this.webClient = webClient;
         this.spotifyAuthService = spotifyAuthService;
+        this.webClient = webClient;
     }
 
     public List<SpotifyTrackDTO> searchMusic(String query) {
@@ -97,25 +79,6 @@ public class MusicPostService {
                 .collect(Collectors.toList());
     }
 
-    @Transactional
-    public List<MusicHistoryDTO> getMusicHistoryByMember(Long memberId) {
-        if (!memberRepository.existsById(memberId)) {
-            throw new MusicException(MusicErrorCode.MEMBER_NOT_FOUND);
-        }
-
-        List<MusicPost> posts = musicPostRepository.findByMemberId(memberId);
-
-        return posts.stream()
-                .map(post -> MusicHistoryDTO.builder()
-                        .musicId(post.getMusic().getId())
-                        .title(post.getMusic().getName())
-                        .artist(post.getMusic().getArtistName())
-                        .albumImage(post.getMusic().getAlbumImgUrl())
-                        .createdAt(post.getCreatedAt())
-                        .build())
-                .collect(Collectors.toList());
-    }
-
     private MusicRequestDTO fetchSpotifyTrackDetail(String spotifyId) {
 
         String token = spotifyAuthService.getAccessToken();
@@ -149,22 +112,11 @@ public class MusicPostService {
                 .build();
     }
 
-
     @Transactional
-    public MusicPost createMusicPost(Long memberId, Long groupId, String spotifyId, String content) {
+    public Music findOrCreateMusic(String spotifyId) {
 
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MusicException(MusicErrorCode.MEMBER_NOT_FOUND));
-
-        Group group = null;
-        if (groupId != null) {
-            group = groupRepository.findById(groupId)
-                    .orElseThrow(() -> new MusicException(MusicErrorCode.GROUP_NOT_FOUND));
-        }
-
-        Music music = musicRepository.findBySpotifyId(spotifyId)
+        return musicRepository.findBySpotifyId(spotifyId)
                 .orElseGet(() -> {
-
                     MusicRequestDTO track = fetchSpotifyTrackDetail(spotifyId);
 
                     Music newMusic = Music.builder()
@@ -182,17 +134,5 @@ public class MusicPostService {
 
                     return musicRepository.save(newMusic);
                 });
-
-        MusicPost post = MusicPost.builder()
-                .member(member)
-                .music(music)
-                .group(group)
-                .isProfile(true)
-                .content(content)
-                .likeCount(0)
-                .commentCount(0)
-                .build();
-
-        return musicPostRepository.save(post);
     }
 }
